Fix/121 game restru match (#188)
* init restru

* fix lint error

* uncomment code

* fix if statement

* GitHub Actions から差分を Push

* init game restru GAME

* lint

* GitHub Actions から差分を Push

* split types and reduce imports

* GitHub Actions から差分を Push

* set component folder and fix index

* GitHub Actions から差分を Push

* init match restru

* GitHub Actions から差分を Push

* paddle fix up

* GitHub Actions から差分を Push

* udpate

* GitHub Actions から差分を Push

---------
